+++ plugins/interfaces/deploymenttest.go
@@ -1,6 +1,9 @@
 package interfaces
 
-import "context"
+import (
+	"context"
+	"time"
+)
 
 // PostDeploymentTest defines a plugin that validates the health of a new deployment by
 // executing a number of requests to it and then executing the
@@ -8,5 +11,5 @@ type PostDeploymentTest interface {
 	Configurable
 
 	// Execute the tests and return an error if the test fails
-	Execute(ctx context.Context) error
+	Execute(ctx context.Context, interval time.Duration) error
 }